+++ sdk/cosmos/azure-cosmos/azure/cosmos/aio/_container.py
@@ -545,7 +545,7 @@ class ContainerProxy:
             initial_headers: Optional[dict[str, str]] = None,
             max_integrated_cache_staleness_in_ms: Optional[int] = None,
             max_item_count: Optional[int] = None,
-            max_degree_of_parallelism: Optional[int] = None,
+            max_concurrency: Optional[int] = None,
             parameters: Optional[list[dict[str, object]]] = None,
             partition_key: PartitionKeyType,
             populate_index_metrics: Optional[bool] = None,
@@ -580,7 +580,7 @@ class ContainerProxy:
             milliseconds. For accounts configured to use the integrated cache, using Session or Eventual consistency,
             responses are guaranteed to be no staler than this value.
         :keyword int max_item_count: Max number of items to be returned in the enumeration operation.
-        :keyword int max_degree_of_parallelism: The maximum number of concurrent operations that run client side
+        :keyword int max_concurrency: The maximum number of concurrent operations that run client side
             during parallel query execution. A positive value limits the number of concurrent operations to the
             set value. If set to 0 (the default), client-side parallelization is disabled and queries run serially.
             If set to -1, the system automatically decides the number of concurrent operations to run.
@@ -643,7 +643,7 @@ class ContainerProxy:
             initial_headers: Optional[dict[str, str]] = None,
             max_integrated_cache_staleness_in_ms: Optional[int] = None,
             max_item_count: Optional[int] = None,
-            max_degree_of_parallelism: Optional[int] = None,
+            max_concurrency: Optional[int] = None,
             parameters: Optional[list[dict[str, object]]] = None,
             populate_index_metrics: Optional[bool] = None,
             populate_query_metrics: Optional[bool] = None,
@@ -678,7 +678,7 @@ class ContainerProxy:
             milliseconds. For accounts configured to use the integrated cache, using Session or Eventual consistency,
             responses are guaranteed to be no staler than this value.
         :keyword int max_item_count: Max number of items to be returned in the enumeration operation.
-        :keyword int max_degree_of_parallelism: The maximum number of concurrent operations that run client side
+        :keyword int max_concurrency: The maximum number of concurrent operations that run client side
             during parallel query execution. A positive value limits the number of concurrent operations to the
             set value. If set to 0 (the default), the system does not impose a limit and queries run serially.
             If set to -1, the system automatically decides the number of concurrent operations to run.
@@ -736,7 +736,7 @@ class ContainerProxy:
             initial_headers: Optional[dict[str, str]] = None,
             max_integrated_cache_staleness_in_ms: Optional[int] = None,
             max_item_count: Optional[int] = None,
-            max_degree_of_parallelism: Optional[int] = None,
+            max_concurrency: Optional[int] = None,
             parameters: Optional[list[dict[str, object]]] = None,
             populate_index_metrics: Optional[bool] = None,
             populate_query_metrics: Optional[bool] = None,
@@ -770,7 +770,7 @@ class ContainerProxy:
             milliseconds. For accounts configured to use the integrated cache, using Session or Eventual consistency,
             responses are guaranteed to be no staler than this value.
         :keyword int max_item_count: Max number of items to be returned in the enumeration operation.
-        :keyword int max_degree_of_parallelism: The maximum number of concurrent operations that run client side
+        :keyword int max_concurrency: The maximum number of concurrent operations that run client side
             during parallel query execution. A positive value limits the number of concurrent operations to the
             set value. If set to 0 (the default), the system does not impose a limit and queries run serially.
             If set to -1, the system automatically decides the number of concurrent operations to run.
@@ -917,8 +917,8 @@ class ContainerProxy:
             feed_options["maxIntegratedCacheStaleness"] = max_integrated_cache_staleness_in_ms
         if utils.valid_key_value_exist(kwargs, "continuation_token_limit"):
             feed_options["responseContinuationTokenLimitInKb"] = kwargs.pop("continuation_token_limit")
-        if utils.valid_key_value_exist(kwargs, "max_degree_of_parallelism"):
-            feed_options["maxDegreeOfParallelism"] = kwargs.pop("max_degree_of_parallelism")
+        if utils.valid_key_value_exist(kwargs, "max_concurrency"):
+            feed_options["maxConcurrency"] = kwargs.pop("max_concurrency")
 
         # populate availability_strategy
         if (Constants.Kwargs.AVAILABILITY_STRATEGY in feed_options